+++ src/firmware/restoreOfficialDialog/actions.ts
@@ -1,0 +1,7 @@
+// SPDX-License-Identifier: MIT
+// Copyright (c) 2022 The Pybricks Authors
+
+export {
+    show as firmwareRestoreOfficialDialogShow,
+    hide as firmwareRestoreOfficialDialogHide,
+} from './redux';